+++ app/src/main/java/org/simple/clinic/home/overdue/search/OverdueSearchScreen.kt
@@ -37,7 +37,6 @@ import org.simple.clinic.databinding.ListItemOverduePlaceholderBinding
 import org.simple.clinic.databinding.ListItemSearchOverdueSelectAllButtonBinding
 import org.simple.clinic.databinding.ScreenOverdueSearchBinding
 import org.simple.clinic.di.injector
-import org.simple.clinic.feature.Feature
 import org.simple.clinic.feature.Features
 import org.simple.clinic.home.overdue.OverdueAppointment
 import org.simple.clinic.home.overdue.search.OverdueSearchProgressState.DONE
@@ -55,6 +54,9 @@ import org.simple.clinic.summary.OpenIntention
 import org.simple.clinic.summary.PatientSummaryScreenKey
 import org.simple.clinic.util.RuntimeNetworkStatus
 import org.simple.clinic.util.UserClock
+import org.simple.clinic.util.applyInsetsBottomPadding
+import org.simple.clinic.util.applyStatusBarPadding
+import org.simple.clinic.util.lightStatusBar
 import org.simple.clinic.util.unsafeLazy
 import org.simple.clinic.widgets.PagingItemAdapter
 import org.simple.clinic.widgets.UiEvent
@@ -96,6 +98,9 @@ class OverdueSearchScreen : BaseScreen<
 
   @Inject
   lateinit var effectHandlerFactory: OverdueSearchEffectHandler.Factory
+
+  private val appbar
+    get() = binding.overdueSearchAppBar
 
   private val overdueSearchToolbar
     get() = binding.overdueSearchToolbar
@@ -213,12 +218,25 @@ class OverdueSearchScreen : BaseScreen<
     context.injector<Injector>().inject(this)
   }
 
+  override fun onResume() {
+    super.onResume()
+    lightStatusBar(enabled = true)
+  }
+
+  override fun onStop() {
+    lightStatusBar(enabled = false)
+    super.onStop()
+  }
+
   override fun bindView(layoutInflater: LayoutInflater, container: ViewGroup?): ScreenOverdueSearchBinding {
     return ScreenOverdueSearchBinding.inflate(layoutInflater, container, false)
   }
 
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
+    appbar.applyStatusBarPadding()
+    downloadAndShareButtonFrame.applyInsetsBottomPadding()
+
     overdueSearchToolbar.setNavigationOnClickListener {
       router.pop()
     }