+++ app/src/main/java/org/simple/clinic/registration/facility/RegistrationFacilitySelectionScreen.kt
@@ -25,6 +25,7 @@ import org.simple.clinic.registration.confirmfacility.ConfirmFacilitySheet
 import org.simple.clinic.registration.register.RegistrationLoadingScreen
 import org.simple.clinic.user.OngoingRegistrationEntry
 import org.simple.clinic.util.extractSuccessful
+import org.simple.clinic.util.lightStatusBar
 import org.simple.clinic.widgets.UiEvent
 import java.util.UUID
 import javax.inject.Inject
@@ -85,6 +86,16 @@ class RegistrationFacilitySelectionScreen : BaseScreen<
   override fun onAttach(context: Context) {
     super.onAttach(context)
     context.injector<Injector>().inject(this)
+  }
+
+  override fun onResume() {
+    super.onResume()
+    lightStatusBar(enabled = true)
+  }
+
+  override fun onStop() {
+    lightStatusBar(enabled = false)
+    super.onStop()
   }
 
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {