+++ app/src/main/java/org/simple/clinic/settings/changelanguage/ChangeLanguageScreen.kt
@@ -22,6 +22,9 @@ import org.simple.clinic.navigation.v2.ScreenKey
 import org.simple.clinic.navigation.v2.fragments.BaseScreen
 import org.simple.clinic.settings.Language
 import org.simple.clinic.settings.changelanguage.ChangeLanguageListItem.Event.ListItemClicked
+import org.simple.clinic.util.applyInsetsBottomMargin
+import org.simple.clinic.util.applyStatusBarPadding
+import org.simple.clinic.util.lightStatusBar
 import org.simple.clinic.widgets.ItemAdapter
 import javax.inject.Inject
 
@@ -38,6 +41,9 @@ class ChangeLanguageScreen : BaseScreen<
 
   @Inject
   lateinit var effectHandlerFactory: ChangeLanguageEffectHandler.Factory
+
+  private val appbar
+    get() = binding.appbar
 
   private val toolbar
     get() = binding.toolbar
@@ -86,8 +92,20 @@ class ChangeLanguageScreen : BaseScreen<
     context.injector<Injector>().inject(this)
   }
 
+  override fun onResume() {
+    super.onResume()
+    lightStatusBar(enabled = true)
+  }
+
+  override fun onStop() {
+    lightStatusBar(enabled = false)
+    super.onStop()
+  }
+
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
+    appbar.applyStatusBarPadding()
+    doneButton.applyInsetsBottomMargin()
 
     setupLanguagesList()
     toolbar.setNavigationOnClickListener { router.pop() }